Moved `auth_config` to `auth`
Make it consistent with `proxy` and `resource_pack`
